Backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
email-validator==2.1.0.post1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
python-jose[cryptography]
passlib[bcrypt]
starlette==0.36.3
typing-extensions==4.9.0
anyio==4.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
fastapi-mail==1.4.1
